requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
psycopg2-binary==2.9.10
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0

pytest~=8.3.4
fastapi~=0.115.8
pydantic~=2.10.6
uvicorn~=0.34.0